refactor: movido para dto/user
